requirements.txt
websockets==14.1
pydantic==2.10.3
aiohttp==3.11.11
requests==2.32.3
PyQt6==6.8.0
ollama
mss
Pillow

